+++ src/test/java/com/seeshiontech/vesdk/RenderProcessTest.java
@@ -1,10 +1,10 @@
-package com.seeshiontech.ve.sdk;
+package com.seeshiontech.vesdk;
 
 import com.alibaba.fastjson.JSON;
 import com.alibaba.fastjson.annotation.JSONField;
-import com.seeshiontech.ve.sdk.exceptions.InvalidLicenseException;
-import com.seeshiontech.ve.sdk.exceptions.NotSupportedTemplateException;
-import com.seeshiontech.ve.sdk.exceptions.RenderException;
+import com.seeshiontech.vesdk.exceptions.InvalidLicenseException;
+import com.seeshiontech.vesdk.exceptions.NotSupportedTemplateException;
+import com.seeshiontech.vesdk.exceptions.RenderException;
 import org.junit.Test;
 
 import java.io.File;