+++ ReleaseNotes.md
@@ -1,6 +1,7 @@
 ve-java-sdk 
 ### version 1.3
-+ 更改包名为 com.seeshiontech.ve.sdk
++ 更改包名为 com.seeshiontech.vesdk
+
 
 ### version 1.2
 + 增加 setDynamicSubFiles 方法,为动态模板设置相关子素材